server/requirements.txt
firebase_admin==6.8.0
Flask==3.1.1
flask_cors==5.0.1
protobuf==6.31.0
python-dotenv==1.1.0
plaid-python==32.0.0
waitress==3.0.2
google-cloud-secret-manager==2.23.3
pandas==2.2.3